+++ dflib-echarts/src/test/java/org/dflib/echarts/BarSeriesTest.java
@@ -13,7 +13,7 @@ public class BarSeriesTest extends GenerateScriptHtmlTest {
         String s2 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScriptHtml("_tid", df2);
         assertTrue(s2.contains("type: 'bar'"), s2);
 
-        String s3 = ECharts.chart().defaultSeriesOpts(SeriesOpts.ofBar()).series("y1").generateScriptHtml("_tid", df2);
+        String s3 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScriptHtml("_tid", df2);
         assertTrue(s3.contains("type: 'bar'"), s3);
     }
 
@@ -23,7 +23,7 @@ public class BarSeriesTest extends GenerateScriptHtmlTest {
         String s2 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScriptHtml("_tid", df2);
         assertFalse(s2.contains("stack: "), s2);
 
-        String s3 = ECharts.chart().defaultSeriesOpts(SeriesOpts.ofBar().stack()).series("y1").generateScriptHtml("_tid", df2);
+        String s3 = ECharts.chart().series(SeriesOpts.ofBar().stack(), "y1").generateScriptHtml("_tid", df2);
         assertTrue(s3.contains("stack: 'total'"), s3);
     }
 }