+++ dflib-echarts/src/test/java/org/dflib/echarts/SeriesTest.java
@@ -57,7 +57,7 @@ public class SeriesTest extends GenerateScriptHtmlTest {
         String s2 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScriptHtml("_tid", df2);
         assertTrue(s2.contains("type: 'bar'"), s2);
 
-        String s3 = ECharts.chart().defaultSeriesOpts(SeriesOpts.ofBar()).series("y1").generateScriptHtml("_tid", df2);
+        String s3 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScriptHtml("_tid", df2);
         assertTrue(s3.contains("type: 'bar'"), s3);
 
         String s4 = ECharts.chart().series(SeriesOpts.ofPie(), "y1").generateScriptHtml("_tid", df2);